move `isManagerNode()` to process, only create metrics ism config on manager node fixes #18830
